requirements.txt
tensorflow>=2.13.0
numpy>=1.19.5
matplotlib>=3.5.1
opencv-python>=4.5.5
scikit-learn>=1.0.2
pandas>=1.4.1
Pillow>=9.0.1
albumentations>=1.1.0
streamlit>=1.8.0
seaborn>=0.11.2